refactor(types): move the retired-field-key tombstone registry off the main barrel (objectui#6527 option B)
Maintainer ruling, 2026-08-28: the registry moves to a dedicated internal
subpath, `@object-ui/types/internal/retired-field-keys`, instead of the
package's main barrel. Importing the barrel eagerly evaluates every other
module it re-exports -- including `spec-report.ts`'s read of
`@objectstack/spec/ui` -- which is what widened an unrelated test's partial
spec mock into a failed suite
(packages/app-shell/src/views/metadata-admin/clientValidation.skew.test.tsx)
under the prior (option A) shape. A subpath import never pulls the barrel in,
so that suite passes again with zero changes to the test or its mock.

Four edits, exactly as ruled:
- Remove the registry's re-exports from the `@object-ui/types` main barrel
  (`packages/types/src/index.ts`).
- Add one dedicated internal subpath to the `exports` map
  (`./internal/retired-field-keys`), and move the registry's source under
  `packages/types/src/internal/` to match -- the package's existing `./zod`
  subpath is the precedent for a subpath's source living in a matching real
  subdirectory rather than a flat file behind a renamed export key.
- Rewrite the 3 consuming imports (MetadataService.ts, MetadataFieldsPage.tsx,
  object-fields-io.ts) to the subpath.
- Re-point the registry's own barrel-wiring pin at the subpath, and add a
  companion pin that the registry is NOT exported from the main barrel, so a
  regression back to option A turns a test red instead of drifting silently.

Nothing else changes: the objectui#6526 formula read-door asymmetry stays
encoded and pinned, the sortOrder defensive verdict stays recorded, and
scripts/check-designer-field-key-parity.mjs is untouched (domain:devx lane).

Changeset unchanged: '@object-ui/types': minor (already correct).
